+++ x-pack/plugins/rule_registry/common/types.ts
@@ -7,10 +7,6 @@
 
 import { estypes } from '@elastic/elasticsearch';
 
-export type PutIndexTemplateRequest = estypes.IndicesPutIndexTemplateRequest & {
-  body?: { composed_of?: string[] };
-};
-
 export interface ClusterPutComponentTemplateBody {
   template: {
     settings: {
