+++ x-pack/plugins/rule_registry/server/index.ts
@@ -8,13 +8,17 @@
 import { PluginInitializerContext } from 'src/core/server';
 import { RuleRegistryPlugin } from './plugin';
 
-export * from './config';
 export type { RuleRegistryPluginSetupContract, RuleRegistryPluginStartContract } from './plugin';
-export type { RacRequestHandlerContext, RacApiRequestHandlerContext } from './types';
-export { RuleDataPluginService } from './rule_data_plugin_service';
-export * from './rule_data_plugin_service/index_options';
-export { RuleDataClient } from './rule_data_client';
-export { IRuleDataClient } from './rule_data_client/types';
+export type {
+  RacRequestHandlerContext,
+  RacApiRequestHandlerContext,
+  AlertTypeWithExecutor,
+} from './types';
+
+export * from './config';
+export * from './rule_data_plugin_service';
+export * from './rule_data_client';
+
 export { getRuleData, RuleExecutorData } from './utils/get_rule_executor_data';
 export { createLifecycleRuleTypeFactory } from './utils/create_lifecycle_rule_type_factory';
 export {
@@ -25,7 +29,6 @@ export {
 } from './utils/create_lifecycle_executor';
 export { createPersistenceRuleTypeFactory } from './utils/create_persistence_rule_type_factory';
 export * from './utils/persistence_types';
-export type { AlertTypeWithExecutor } from './types';
 
 export const plugin = (initContext: PluginInitializerContext) =>
   new RuleRegistryPlugin(initContext);
