+++ x-pack/plugins/rule_registry/server/rule_data_plugin_service/errors.ts
@@ -5,7 +5,6 @@
  * 2.0.
  */
 
-/** @deprecated */
 export class RuleDataWriteDisabledError extends Error {
   constructor(message?: string) {
     super(message);
