+++ x-pack/plugins/rule_registry/server/utils/create_lifecycle_executor.test.ts
@@ -27,7 +27,7 @@ import {
   ALERT_RULE_CONSUMER,
   SPACE_IDS,
 } from '../../common/technical_rule_data_field_names';
-import { createRuleDataClientMock } from '../rule_data_client/create_rule_data_client_mock';
+import { createRuleDataClientMock } from '../rule_data_client/rule_data_client.mock';
 import { createLifecycleExecutor } from './create_lifecycle_executor';
 
 describe('createLifecycleExecutor', () => {
