+++ x-pack/plugins/security_solution/server/lib/detection_engine/rule_types/query/create_query_alert_type.test.ts
@@ -36,7 +36,6 @@ describe('Custom query alerts', () => {
     const { services, dependencies, executor } = createRuleTypeMocks();
     const queryAlertType = createQueryAlertType({
       experimentalFeatures: allowedExperimentalValues,
-      indexAlias: 'alerts.security-alerts',
       lists: dependencies.lists,
       logger: dependencies.logger,
       mergeStrategy: 'allFields',
@@ -84,7 +83,6 @@ describe('Custom query alerts', () => {
     const { services, dependencies, executor } = createRuleTypeMocks();
     const queryAlertType = createQueryAlertType({
       experimentalFeatures: allowedExperimentalValues,
-      indexAlias: 'alerts.security-alerts',
       lists: dependencies.lists,
       logger: dependencies.logger,
       mergeStrategy: 'allFields',
